refactor(bookmark): 将文章内容和翻译功能迁移到独立的ArticleRepository
重构书签相关功能，将文章内容获取、缓存和翻译逻辑从BookmarkRepository迁移到新创建的ArticleRepository
简化BookmarkRepository职责，使其只关注书签核心业务
更新相关依赖注入和调用点以适配新的结构
